corpus/benign/mcp/gh-youcanautomate-yca-ai-driven-mobile-automation/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "appium-mcp"
version = "1.0.14"
description = "AI-driven mobile test automation framework with Appium MCP server, Page Object Model, and AWS Bedrock integration - Write test automation in natural language with automatic page objects and test generation"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "You Can Automate", email = "youcanautomate@gmail.com"}
]
maintainers = [
    {name = "You Can Automate", email = "youcanautomate@gmail.com"}
]
keywords = [
    "appium",
    "mobile-automation",
    "mcp",
    "model-context-protocol",
    "ai",
    "bedrock",
    "testing",
    "ios",
    "android",
    "test-automation",
    "page-object-model"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.11.0",
    "appium-python-client>=3.0.0",
    "pyyaml>=6.0.3",
    "python-dotenv>=1.0.0",
    "boto3>=1.26.0",
    "rich>=13.0.0",
    "pytest>=7.0.0",
    "selenium>=4.0.0",
    "jinja2>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "build>=0.10.0",
    "twine>=4.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/youcanautomate-yca/ai-driven-mobile-automation"
Documentation = "https://github.com/youcanautomate-yca/ai-driven-mobile-automation/blob/main/README.md"
Installation = "https://github.com/youcanautomate-yca/ai-driven-mobile-automation/blob/main/INSTALLATION.md"
Repository = "https://github.com/youcanautomate-yca/ai-driven-mobile-automation.git"
Issues = "https://github.com/youcanautomate-yca/ai-driven-mobile-automation/issues"
Changelog = "https://github.com/youcanautomate-yca/ai-driven-mobile-automation/releases"
"Bug Tracker" = "https://github.com/youcanautomate-yca/ai-driven-mobile-automation/issues"

[project.scripts]
appium-mcp-server = "server:main"
appium-mcp-run-yaml = "run_yaml:main"
appium-mcp-generate-tests = "generate_test_from_workflow:generate_test_from_workflow"
appium-mcp-generate-pages = "generate_page_objects:generate_page_objects"
appium-mcp-generate-regression = "generate_regression_test:generate_regression_test"
appium-mcp-chatbot = "chatbot:main"

[tool.setuptools]
py-modules = [
    "server",
    "command",
    "orchestrator",
    "session_store",
    "config",
    "logger",
    "bedrock_client",
    "mcp_client",
    "run_yaml",
    "generate_test_from_workflow",
    "generate_page_objects",
    "generate_test_with_page_objects",
    "generate_regression_test",
    "chatbot",
    "tools_session",
    "tools_interactions",
    "tools_navigations",
    "tools_app_management",
    "tools_context",
    "tools_ios",
    "tools_documentation",
    "tools_test_generation",
    "debug_app_elements",
    "inspect_app_elements",
    "locator_validator",
    "locator_validation_service",
    "chatbot_page_object_orchestrator",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["page_objects*", "prompts*", "templates*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"
